sort VerletClusterCells files in separate directory
