+++ src/autopas/containers/verletClusterCells/VerletClusterCells.h
@@ -14,8 +14,8 @@
 #include "autopas/containers/CellBorderAndFlagManager.h"
 #include "autopas/containers/ParticleContainer.h"
 #include "autopas/containers/cellPairTraversals/CellPairTraversal.h"
-#include "autopas/containers/verletClusterLists/VerletClusterCellsParticleIterator.h"
-#include "autopas/containers/verletClusterLists/traversals/VerletClusterTraversalInterface.h"
+#include "autopas/containers/verletClusterCells/VerletClusterCellsParticleIterator.h"
+#include "autopas/containers/verletClusterCells/traversals/VerletClusterCellsTraversalInterface.h"
 #include "autopas/iterators/ParticleIterator.h"
 #include "autopas/iterators/RegionParticleIterator.h"
 #include "autopas/utils/ArrayMath.h"
@@ -65,7 +65,7 @@ class VerletClusterCells : public ParticleContainer<FullParticleCell<Particle>> 
    * @param traversal to be used used
    */
   void iteratePairwise(TraversalInterface *traversal) override {
-    auto *traversalInterface = dynamic_cast<VerletClusterTraversalInterface<FullParticleCell<Particle>> *>(traversal);
+    auto *traversalInterface = dynamic_cast<VerletClusterCellsTraversalInterface<FullParticleCell<Particle>> *>(traversal);
     auto *cellPairTraversal = dynamic_cast<CellPairTraversal<FullParticleCell<Particle>> *>(traversal);
 
     if ((!traversalInterface) or (!cellPairTraversal)) {
@@ -141,7 +141,7 @@ class VerletClusterCells : public ParticleContainer<FullParticleCell<Particle>> 
    * @param traversal The used traversal.
    */
   void rebuildNeighborLists(TraversalInterface *traversal) override {
-    auto *traversalInterface = dynamic_cast<VerletClusterTraversalInterface<FullParticleCell<Particle>> *>(traversal);
+    auto *traversalInterface = dynamic_cast<VerletClusterCellsTraversalInterface<FullParticleCell<Particle>> *>(traversal);
     if (!traversalInterface) {
       autopas::utils::ExceptionHandler::exception(
           "trying to use a traversal of wrong type in VerletClusterCells::iteratePairwise");
